integracion de back (coworkings) con filtros
