+++ front/src/app/coworkings/page.ts
@@ -1,1 +1,1 @@
-export { default } from "@/components/coworkings";
+export { default } from "./CoworksRoute";
