--- front/src/app/coworks/page.ts
@@ -1,1 +1,0 @@
-export { default } from "./CoworksRoute";
