+++ front/src/components/CoworkCard/index.tsx
@@ -1,49 +1,52 @@
+import Link from "next/link";
 import "./style.css"
 export const CoworkCard =({cowork})=>{
 
     return(
         <div className="property-card">
             <figure className="card-banner img-holder" style={{width: "800", height: "533"}}>
-                <img src={cowork.image} width="800" height="533" loading="lazy"
+                <img src={cowork.thumbnail} width="800" height="533" loading="lazy"
                 alt="espacio 5" className="img-cover"/>
             </figure>
             {/* <button className="card-action-btn favourite-btn " aria-label="add to favourite">
                 <div className="heart-container" title="Like">
-            <input type="checkbox" className="checkbox" id="Give-It-An-Id"/>
-            <div className="svg-container">
-                <svg viewBox="0 0 24 24" className="svg-outline">
-                    <path d="M17.5,1.917a6.4,6.4,0,0,0-5.5,3.3,6.4,6.4,0,0,0-5.5-3.3A6.8,6.8,0,0,0,0,8.967c0,4.547,4.786,9.513,8.8,12.88a4.974,4.974,0,0,0,6.4,0C19.214,18.48,24,13.514,24,8.967A6.8,6.8,0,0,0,17.5,1.917Zm-3.585,18.4a2.973,2.973,0,0,1-3.83,0C4.947,16.006,2,11.87,2,8.967a4.8,4.8,0,0,1,4.5-5.05A4.8,4.8,0,0,1,11,8.967a1,1,0,0,0,2,0,4.8,4.8,0,0,1,4.5-5.05A4.8,4.8,0,0,1,22,8.967C22,11.87,19.053,16.006,13.915,20.313Z">
-                    </path>
-                </svg>
-                <svg viewBox="0 0 24 24" className="svg-filled">
-                    <path d="M17.5,1.917a6.4,6.4,0,0,0-5.5,3.3,6.4,6.4,0,0,0-5.5-3.3A6.8,6.8,0,0,0,0,8.967c0,4.547,4.786,9.513,8.8,12.88a4.974,4.974,0,0,0,6.4,0C19.214,18.48,24,13.514,24,8.967A6.8,6.8,0,0,0,17.5,1.917Z">
-                    </path>
-                </svg>
-                <svg className="svg-celebrate" width="100" height="100">
-                    <polygon points="10,10 20,20"></polygon>
-                    <polygon points="10,50 20,50"></polygon>
-                    <polygon points="20,80 30,70"></polygon>
-                    <polygon points="90,10 80,20"></polygon>
-                    <polygon points="90,50 80,50"></polygon>
-                    <polygon points="80,80 70,70"></polygon>
-                </svg>
-            </div>
-        </div>
+                    <input type="checkbox" className="checkbox" id="Give-It-An-Id"/>
+                    <div className="svg-container">
+                        <svg viewBox="0 0 24 24" className="svg-outline">
+                            <path d="M17.5,1.917a6.4,6.4,0,0,0-5.5,3.3,6.4,6.4,0,0,0-5.5-3.3A6.8,6.8,0,0,0,0,8.967c0,4.547,4.786,9.513,8.8,12.88a4.974,4.974,0,0,0,6.4,0C19.214,18.48,24,13.514,24,8.967A6.8,6.8,0,0,0,17.5,1.917Zm-3.585,18.4a2.973,2.973,0,0,1-3.83,0C4.947,16.006,2,11.87,2,8.967a4.8,4.8,0,0,1,4.5-5.05A4.8,4.8,0,0,1,11,8.967a1,1,0,0,0,2,0,4.8,4.8,0,0,1,4.5-5.05A4.8,4.8,0,0,1,22,8.967C22,11.87,19.053,16.006,13.915,20.313Z">
+                            </path>
+                        </svg>
+                        <svg viewBox="0 0 24 24" className="svg-filled">
+                            <path d="M17.5,1.917a6.4,6.4,0,0,0-5.5,3.3,6.4,6.4,0,0,0-5.5-3.3A6.8,6.8,0,0,0,0,8.967c0,4.547,4.786,9.513,8.8,12.88a4.974,4.974,0,0,0,6.4,0C19.214,18.48,24,13.514,24,8.967A6.8,6.8,0,0,0,17.5,1.917Z">
+                            </path>
+                        </svg>
+                        <svg className="svg-celebrate" width="100" height="100">
+                            <polygon points="10,10 20,20"></polygon>
+                            <polygon points="10,50 20,50"></polygon>
+                            <polygon points="20,80 30,70"></polygon>
+                            <polygon points="90,10 80,20"></polygon>
+                            <polygon points="90,50 80,50"></polygon>
+                            <polygon points="80,80 70,70"></polygon>
+                        </svg>
+                    </div>
+                </div>
             </button> */}
             <div className="card-content">
-                <h3 className="h3">
-                  <a href="#" className="card-title">{cowork.name}</a>
-                </h3>
+                <Link href={`/coworkings/${cowork.id}`} >
+                    <h3 className="h3">
+                     {cowork.name}
+                    </h3>
+                </Link>
                 <ul className="card-list">
                     <li className="card-item">
                     <svg viewBox="0 0 24 24" strokeLinecap="round" strokeLinejoin="round"><circle cx="12" cy="12" r="10"></circle><polyline points="12 6 12 12 16 14"></polyline></svg>
                     </li>
                     <li className="card-item">
-                        <span className="item-text">{cowork.dateOpen}am</span>
+                        <span className="item-text">{cowork.open}am</span>
                     </li>
                     <span>-</span>
                     <li className="card-item">
-                        <span className="item-text">{cowork.dateClose}pm</span>
+                        <span className="item-text">{cowork.close}pm</span>
                     </li>
                 </ul>
                 <div className="card-meta">
@@ -53,7 +56,7 @@ export const CoworkCard =({cowork})=>{
                             Ubicacion
                         </span>
                         <span className="meta-text">
-                          {`${cowork.location.city}, ${cowork.location.state}, ${cowork.location.country}`}
+                          {`${cowork.city}, ${cowork.state}, ${cowork.country}`}
                         </span>
                     </div>
                     {/* <div>
