+++ front/src/components/coworkings/style.css
@@ -1,0 +1,259 @@
+:root {
+
+    --oxford-blue_60: hsla(224, 34%, 13%, 0.6);
+    --green-pigment: hsl(142, 76%, 36%);
+    --oxford-blue: hsl(224, 34%, 13%);
+    --maximum-red: hsl(0, 72%, 51%);
+    --eerie-black: hsl(0, 0%, 9%);
+    --mint-cream: hsl(146, 41%, 97%);
+    --cadet-gray: hsl(215, 20%, 65%);
+    --light-gray: hsl(216, 12%, 84%);
+    --carcoal_20: hsla(214, 19%, 29%, 0.2);
+    --carcoal_15: hsla(214, 19%, 29%, 0.15);
+    --carcoal_10: hsla(214, 19%, 29%, 0.10);
+    --sea-green: hsl(142, 72%, 29%);
+    --gainsboro: hsl(220, 13%, 91%);
+    --white_70: hsla(0, 0%, 100%, 0.7);
+    --gunmetal: hsl(215, 28%, 17%);
+    --manatee: hsl(218, 11%, 65%);
+    --mango: hsl(43, 96%, 56%);
+    --white: hsl(0, 0%, 100%);
+    --black: hsl(0, 0%, 0%);
+  
+    /**
+     * typography
+     */
+  
+    --ff-spartan: 'League Spartan', sans-serif;
+  
+    --fs-1: 3.6rem;
+    --fs-2: 2.4rem;
+    --fs-3: 2rem;
+    --fs-4: 1.8rem;
+    --fs-5: 1.7rem;
+  
+    --fw-600: 600;
+    --fw-500: 500;
+  
+    /**
+     * spacing
+     */
+  
+    --section-padding: 35px;
+  
+    /**
+     * shadow
+     */
+  
+    --shadow-1: 0 1px 2px hsla(0, 0%, 0%, 0.2);
+    --shadow-2: 0 10px 15px -5px var(--carcoal_20);
+    --shadow-3: 0 10px 25px -3px var(--carcoal_20);
+    --shadow-4: 0 0 3px var(--carcoal_15);
+    --shadow-5: 0 20px 25px -5px var(--carcoal_10),
+                0 10px 10px -5px var(--carcoal_10);
+  
+    /**
+     * border radius
+     */
+  
+    --radius-16: 16px;
+    --radius-12: 12px;
+    --radius-6: 6px;
+    --radius-4: 4px;
+  
+    /**
+     * transition
+     */
+  
+    --transition: 0.25s ease;
+    --cubic-out: cubic-bezier(0.33, 0.85, 0.4, 0.96);
+  
+  }
+  
+  
+  /*-----------------------------------*\
+    #REUSED STYLE
+  \*-----------------------------------*/
+  
+  .btn {
+    background-color: var(--green-pigment);
+    color: var(--white);
+    font-weight: var(--fw-500);
+    max-width: max-content;
+    padding: var(--padding, 12px 28px);
+  }
+  
+  .btn:is(:hover, :focus) { background-color: var(--sea-green); }
+  
+  /* .btn-secondary { border-radius: 50px; } */
+  
+  .btn-primary { border-radius: 16px; }
+  
+  .h1 { font-size: var(--fs-1); }
+  
+  .h2,
+  .h3 {
+    color: var(--oxford-blue);
+    line-height: 1.5;
+  }
+  
+  .h2 {
+    font-size: var(--fs-2);
+    font-weight: var(--fw-500);
+  }
+  
+  .h3 {
+    font-size: var(--fs-3);
+    font-weight: var(--fw-500);
+  }
+  
+  .img-holder {
+    aspect-ratio: var(--width) / var(--height);
+    background-color: var(--gainsboro);
+  }
+  
+  .img-cover {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+  
+  .section-title { color: var(--oxford-blue); }
+  
+  .section-text { line-height: 1.6; }
+  
+  
+  /*-----------------------------------*\
+    #PROPERTY
+  \*-----------------------------------*/
+  
+  .property{
+    margin-top: 5rem;
+    margin-bottom: 8rem;
+  }
+
+  .property :is(.section-title, .section-text) { text-align: center; }
+  
+  .property .section-text { margin-block: 15px 60px; }
+  
+  .property-list {
+    display: grid;
+    gap: 30px;
+  }
+  
+  /*-----------------------------------*\
+    #MEDIA QUERIES
+  \*-----------------------------------*/
+  
+  /**
+   * responsive for larger than 640px screen
+   */
+  
+  @media (min-width: 640px) {
+  
+    /**
+     * REUSED STYLE
+     */
+  
+    .container {
+      max-width: 600px;
+      width: 100%;
+      margin-inline: auto;
+    }
+  
+  }
+  
+  
+  
+  
+  
+  /**
+   * responsive for larger than 768px screen
+   */
+  
+  @media (min-width: 750px) {
+  
+    /**
+     * CUSTOM PROPERTY
+     */
+  
+    :root {
+  
+      /**
+       * typography
+       */
+  
+      --fs-2: 3rem;
+  
+    }
+  
+  
+  
+    /**
+     * REUSED STYLE
+     */
+  
+    .container,
+    .hero-form-wrapper { max-width: 780px; }
+
+    /**
+     *  PROPERTY
+     */
+  
+    .property-list { grid-template-columns: 1fr 1fr; }
+  
+  }
+  
+
+  
+  @media (min-width: 920px) {
+  
+    /**
+     * CUSTOM PROPERTY
+     */
+  
+    :root {
+  
+      /**
+       * typography
+       */
+  
+      --fs-1: 4.8rem;
+  
+      /**
+       * spacing
+       */
+  
+      --section-padding: 40px;
+  
+    }
+  
+  
+  
+    /**
+     * REUSED STYLE
+     */
+  
+    .container,
+
+    .property-list { grid-template-columns: repeat(3, 1fr); }
+
+  }
+  
+  
+  
+  
+  
+  /**
+   * responsive for larger than 1200px screen
+   */
+  
+  @media (min-width: 1200px) {
+  
+    /**
+     * REUSED STYLE
+     */
+  
+    .container{ max-width: 1120px; }
+
+  }
