+++ front/src/service/FilterCoworks/index.tsx
@@ -1,44 +1,41 @@
 export const FilterCoworks = (coworks: any[], filter: any) => {
-    const toFilter = coworks;
-    const filterElement = filter;
-  
-    const funcionfiltrar = () => {
-      return toFilter.filter(element => {
-        // Comprobar la disponibilidad
-        if (filterElement.aviable.length > 0) {
-          if (filterElement.aviable.includes('aviable') && !element.aviability) {
-            return false;
-          }
-          if (filterElement.aviable.includes('not-aviable') && element.aviability) {
-            return false;
-          }
-        }
-        // Comprobar la valoración
-        if (filterElement.rating.length > 0 && !filterElement.rating.includes(String(element.rate))) {
-          return false;
-        }
-        // Comprobar la ubicación
-        if (filterElement.location.country && filterElement.location.country !== element.location.country) {
-          return false;
-        }
-        if (filterElement.location.state && filterElement.location.state !== element.location.state) {
-          return false;
-        }
-        if (filterElement.location.city && filterElement.location.city !== element.location.city) {
-          return false;
-        }
-        // Si pasa todos los filtros, incluir el elemento
-        return true;
-      });
-    }
-  
-    const filteredCoworks = funcionfiltrar();
-  
-    return filteredCoworks;
-  }
-  
-  
-  
+  const toFilter = coworks;
+  const filterElement = filter;
 
-  
-  export default FilterCoworks;
+  const funcionfiltrar = () => {
+    return toFilter.filter(element => {
+      // Solo incluir los coworkings con status 'active'
+      if (element.status !== 'active') {
+        return false;
+      }
+      // Comprobar la ubicación
+      if (filterElement.location.country && filterElement.location.country !== element.country) {
+        return false;
+      }
+      if (filterElement.location.state && filterElement.location.state !== element.state) {
+        return false;
+      }
+      if (filterElement.location.city && filterElement.location.city !== element.city) {
+        return false;
+      }
+      // Comprobar la capacidad
+      if (filterElement.capacity.length > 0) {
+        const capacity = element.capacity;
+        if (!filterElement.capacity.some(range => {
+          const [min, max] = range.split('-').map(Number);
+          return capacity >= min && capacity <= max;
+        })) {
+          return false;
+        }
+      }
+      // Si pasa todos los filtros, incluir el elemento
+      return true;
+    });
+  }
+
+  const filteredCoworks = funcionfiltrar();
+
+  return filteredCoworks;
+}
+
+export default FilterCoworks;
