+++ front/src/service/getAllCoworks/arrayExample.ts
@@ -1,145 +1,87 @@
 export const arrayExample = [
     {
-        id: 1,
-        name: "Creative Hub",
-        dateOpen: "07:00",
-        dateClose: "10:00",
-        location: { country: "Argentina", state: "Cordoba", city: "Fuente Palmera" },
-        rate: 5,
-        email: "creativehub@gmail.com",
-        contactPhone: 80284352,
-        spaces: 3,
-        aviability: true,
-        image: "https://wearecloudworks.com/wp-content/uploads/2022/01/beneficios-coworking.jpeg"
+      name: "Coworking 1",
+      phone: "123456789",
+      email: "coworking1@mail.com",
+      open: "08:00",
+      close: "20:00",
+      address: "Calle 1, 123",
+      country: "Agentina",
+      state: "Buenos Aires",
+      city: "La Plata",
+      capacity: 20,
+      status: "active",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/gettyimages-1161987904-640x640_gtz1ok.jpg"
     },
     {
-        id: 2,
-        name: "TechSpace",
-        dateOpen: "08:00",
-        dateClose: "18:00",
-        location: { country: "Argentina", state: "Buenos Aires", city: "Palermo" },
-        rate: 5,
-        email: "techspace@gmail.com",
-        contactPhone: 81234567,
-        spaces: 5,
-        aviability: false,
-        image: "https://plus.unsplash.com/premium_photo-1684769161054-2fa9a998dcb6?q=80&w=1000&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MXx8Y293b3JraW5nfGVufDB8fDB8fHww"
+      name: "Coworking 2",
+      phone: "987654321",
+      email: "coworking2@mail.com",
+      open: "09:00",
+      close: "21:00",
+      address: "Calle 2, 456",
+      country: "Colombia",
+      state: "Antioquia",
+      city: "Antioquia",
+      capacity: 15,
+      status: "active",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/istockphoto-1357829448-640x640_qe3nqo.jpg"
     },
     {
-        id: 3,
-        name: "Innovative Office",
-        dateOpen: "09:00",
-        dateClose: "17:00",
-        location: { country: "Argentina", state: "Rosario", city: "Centro" },
-        rate: 4,
-        email: "innovativeoffice@gmail.com",
-        contactPhone: 82345678,
-        spaces: 8,
-        aviability: true,
-        image: "https://media.licdn.com/dms/image/D5612AQFVPmOhEk6aEg/article-cover_image-shrink_720_1280/0/1678343184566?e=2147483647&v=beta&t=JhnF0D9n2RYYyRakH3HXxMf61f9yKIbQCBU9N2J3NAE"
+      name: "Coworking 3",
+      phone: "555555555",
+      email: "coworking3@mail.com",
+      open: "07:30",
+      close: "19:30",
+      address: "Calle 3, 789",
+      country: "Argentina",
+      state: "Tucuman",
+      city: "Tucuman",
+      capacity: 25,
+      status: "pending",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/istockphoto-1279090224-640x640_t2u9zl.jpg"
     },
     {
-        id: 4,
-        name: "CoWorkSpot",
-        dateOpen: "08:30",
-        dateClose: "19:00",
-        location: { country: "Argentina", state: "Mendoza", city: "Ciudad" },
-        rate: 3,
-        email: "coworkspot@gmail.com",
-        contactPhone: 83456789,
-        spaces: 6,
-        aviability: true,
-        image: "https://media.licdn.com/dms/image/D5612AQGNFj_iZIRQ6g/article-cover_image-shrink_720_1280/0/1693636697937?e=2147483647&v=beta&t=kWS-LAgE175X4GOQ6V2CKSgdm1t6fFeNXVYVyM-RE8k"
+      name: "Coworking 4",
+      phone: "999999999",
+      email: "coworking4@mail.com",
+      open: "08:30",
+      close: "20:30",
+      address: "Calle 4, 101112",
+      country: "Colombia",
+      state: "Bolivar",
+      city: "Arjona",
+      capacity: 18,
+      status: "active",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/gettyimages-1214699141-640x640_bhzal2.jpg"
     },
     {
-        id: 5,
-        name: "Urban Workspace",
-        dateOpen: "06:00",
-        dateClose: "15:00",
-        location: { country: "Argentina", state: "Salta", city: "Centro" },
-        rate: 2,
-        email: "urbanworkspace@gmail.com",
-        contactPhone: 84567890,
-        spaces: 4,
-        aviability: false,
-        image: "https://img.cofynd.com/images/original/3c0fd74f882a8afa1705f33f9459c133bf286242.jpg"
+      name: "Coworking 5",
+      phone: "777777777",
+      email: "coworking5@mail.com",
+      open: "07:45",
+      close: "19:45",
+      address: "Calle 5, 131415",
+      country: "Argentina",
+      state: "Tucuman",
+      city: "Tucuman",
+      capacity: 22,
+      status: "pending",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/gettyimages-1157185547-640x640_qrhflf.jpg"
     },
     {
-        id: 6,
-        name: "Digital Nexus",
-        dateOpen: "09:30",
-        dateClose: "17:30",
-        location: { country: "Mexico", state: "Mexico City", city: "Condesa" },
-        rate: 4,
-        email: "digitalnexus@gmail.com",
-        contactPhone: 85678901,
-        spaces: 7,
-        aviability: false,
-        image: "https://qph.cf2.quoracdn.net/main-qimg-d3f1a883957f1fd0242b43e762e5f4a5-lq"
-    },
-    {
-        id: 7,
-        name: "Work Café",
-        dateOpen: "07:30",
-        dateClose: "16:30",
-        location: { country: "Chile", state: "Santiago", city: "Providencia" },
-        rate: 5,
-        email: "workcafe@gmail.com",
-        contactPhone: 86789012,
-        spaces: 10,
-        aviability: true,
-        image: "https://images.adsttc.com/media/images/5e34/195f/3312/fd12/f700/0163/newsletter/21a.jpg?1580472661"
-    },
-    {
-        id: 8,
-        name: "InnovateSpace",
-        dateOpen: "08:00",
-        dateClose: "18:00",
-        location: { country: "Peru", state: "Lima", city: "Miraflores" },
-        rate: 3,
-        email: "innovatespace@gmail.com",
-        contactPhone: 87890123,
-        spaces: 9,
-        aviability: false,
-        image: "https://i.pinimg.com/736x/7f/6e/eb/7f6eeb959463ac50c3be066293aef9ef.jpg"
-    },
-    {
-        id: 9,
-        name: "CollabLab",
-        dateOpen: "09:00",
-        dateClose: "17:00",
-        location: { country: "Colombia", state: "Bogota", city: "Chapinero" },
-        rate: 2,
-        email: "collablab@gmail.com",
-        contactPhone: 88901234,
-        spaces: 12,
-        aviability: true,
-        image: "https://wearegroundwork.com/wp/wp-content/uploads/blog-10-beneficios-coworking.jpg"
-    },
-    {
-        id: 10,
-        name: "SmartSpace",
-        dateOpen: "07:00",
-        dateClose: "15:00",
-        location: { country: "Argentina", state: "La Plata", city: "Centro" },
-        rate: 4,
-        email: "smartspace@gmail.com",
-        contactPhone: 89012345,
-        spaces: 6,
-        aviability: false,
-        image: "https://coworkidea.com/wp-content/uploads/2020/02/CoworkIdea_Imagen02.jpg"
-    },
-    {
-        id: 11,
-        name: "Creative Hub",
-        dateOpen: "07:00",
-        dateClose: "10:00",
-        location: { country: "Argentina", state: "Cordoba", city: "Fuente Palmera" },
-        rate: 5,
-        email: "creativehub@gmail.com",
-        contactPhone: 80284352,
-        spaces: 3,
-        aviability: false,
-        image: "https://www.ceupe.do/images/easyblog_articles/298/coworking.jpg"
-    },
-];
+      name: "Coworking 6",
+      phone: "333333333",
+      email: "coworking6@mail.com",
+      open: "09:30",
+      close: "21:30",
+      address: "Calle 6, 161718",
+      country: "Argentina",
+      state: "Cordoba",
+      city: "Cordoba",
+      capacity: 17,
+      status: "inactive",
+      thumbnail: "https://res.cloudinary.com/drwhxgsfu/image/upload/v1715195174/coworkings_coco%2B/gettyimages-1157235072-640x640_uspmdc.jpg"
+    }
+  ]
+
