+++ front/src/service/getAllCoworks/index.tsx
@@ -1,16 +1,20 @@
+const urlBase = process.env.NEXT_PUBLIC_API_URL;
 import { ICowork } from "./types";
-import { arrayExample } from "./arrayExample";
 
-// Define la función getAllCoworks para que devuelva directamente el arrayExample
-export const getAllCoworks = () => {
-    try {
-        return arrayExample;
-
-        
-    } catch (error) {
-        console.error("Error fetching products:", error);
-        return [];
+export const getAllCoworks = async (): Promise<ICowork[]> => {
+  try {
+    const response = await fetch(`http://localhost:3000/coworkings`);
+    
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
     }
+
+    const data: ICowork[] = await response.json();
+    return data;
+  } catch (error) {
+    console.error("Error fetching products:", error);
+    return [];
+  }
 };
 
 export default getAllCoworks;
